Change manager directory structure
