+++ pygeppetto/managers/__init__.py
@@ -1,0 +1,3 @@
+from .geppetto_manager import GeppettoManager
+
+__all__ = ['GeppettoManager']
