+++ pygeppetto/managers/download_manager.py
@@ -1,0 +1,4 @@
+
+
+class DataSourceManager(object):
+    pass
