+++ pygeppetto/managers/experiment_manager.py
@@ -1,0 +1,4 @@
+
+
+class ExperimentManager(object):
+    pass
