+++ pygeppetto/managers/experiment_run_manager.py
@@ -1,0 +1,6 @@
+
+
+class ExperimentRunManager(object):
+    @classmethod
+    def queueExperiment(cls, user, experiment):
+        pass
