+++ pygeppetto/managers/project_manager.py
@@ -1,0 +1,3 @@
+
+class ProjectManager(object):
+    pass
